[FEAT] : 다른 유저의 캐릭터 확인후 해당 캐릭터로 생성
