+++ Project/Assets/03.Code/Scripts/LogInSystem/CharaterSelectPage.cs
@@ -1,6 +1,3 @@
-using System.Collections;
-using System.Collections.Generic;
-using System.Diagnostics;
 using UnityEngine;
 using UnityEngine.UI;
