+++ Project/Assets/03.Code/Scripts/LogInSystem/LogInUserData.cs
@@ -1,8 +1,3 @@
-using Newtonsoft.Json;
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
 [System.Serializable]
 public class LogInUserData
 {
