+++ Project/Assets/03.Code/Scripts/PlayerMove/CameraMovement.cs
@@ -1,4 +1,3 @@
-using System;
 using UnityEngine;
 using UnityEngine.InputSystem;
