+++ Project/Assets/04.Level/Scenes/DES/NavMesh-Terrain.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5629ccc8686ac1a4d9c65703ef1d5c4d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 23800000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
