+++ Project/Assets/04.Level/Scenes/GameObject.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a7a1c716a042d9349b6ad635977392d8
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
